# public/dist/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "jwtApp"
  },
  "basePath": "/jwtApp/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "tags": [
    {
      "name": "admin",
      "description": "Admin Apis"
    }
  ],
  "paths": {

    "/admin/signup": {
      "post": {
        "tags": [
          "admin"
        ],
        "description": "Admin Signup API",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Name of Admin",
            "required": false,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Email of Admin",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Password of Admin",
            "required": true,
            "type": "string"
          },
          {
            "name": "age",
            "in": "formData",
            "description": "Age of Admin",
            "required": true,
            "type": "number"
          },
          {
            "name": "platform",
            "in": "header",
            "description": "device OS 1- iOS, 2- Android, 3 - WEB",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/login": {
      "post": {
        "tags": [
          "admin"
        ],
        "description": "Admin Login API",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "Email of Admin",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Password of Admin",
            "required": true,
            "type": "string"
          },
          {
            "name": "platform",
            "in": "header",
            "description": "device OS 1- iOS, 2- Android, 3 - WEB",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/getAdminList": {
      "post": {
        "tags": [
          "admin"
        ],
        "description": "Admin Login API",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "platform",
            "in": "header",
            "description": "device OS 1- iOS, 2- Android, 3 - WEB",
            "required": true,
            "type": "string"
          },
          {
            "name": "accessToken",
            "in": "header",
            "description": "Enter Access Token",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/admin/logout": {
      "post": {
        "tags": [
          "admin"
        ],
        "description": "Admin Logout API",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "platform",
            "in": "header",
            "description": "device OS 1- iOS, 2- Android, 3 - WEB",
            "required": true,
            "type": "string"
          },
          {
            "name": "accessToken",
            "in": "header",
            "description": "Enter Access Token",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/swagger": {
      "x-swagger-pipe": "swagger_raw"
    }
  },
  "definitions": {
    "HelloWorldResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}